add some ui changes

+++ app/src/main/java/com/example/qweather/data/network/WeatherApiService.kt
@@ -12,5 +12,5 @@ interface WeatherApiService {
     suspend fun getCities(): WeatherApiResultResponse<CityResultResponse>
 
     @POST("current_weather/city")
-    suspend fun getForecast(@Query("city_id") cityId: Long): WeatherApiResultResponse<ForecastResult>
+    suspend fun getForecast(@Query("city_id") cityId: Int): WeatherApiResultResponse<ForecastResult>
 }
